pyproject.toml
[project]
name = "kimura-neutral-theory-gaia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "huangyuan", email = "huangyuan@dp.tech" }
]
requires-python = ">=3.12"
dependencies = [
    "gaia-lang>=0.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kimura_neutral_theory"]

[tool.gaia]
type = "knowledge-package"
uuid = "59dbb5af-0cd3-4c04-8b87-72487184bbb8"
